Still working on fixing the rebase, refactored user models
