+++ src/main/java/com/skills/skills/SkillsApplication.java
@@ -7,8 +7,7 @@ import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 @SpringBootApplication
 public class SkillsApplication {
 
-	public static void main(String[] args) {
-		SpringApplication.run(SkillsApplication.class, args);
+	public static void main(String[] args) {SpringApplication.run(SkillsApplication.class, args);
 	}
 
 }
