+++ src/main/java/com/skills/skills/controllers/AuthenticationController.java
@@ -2,25 +2,23 @@ package com.skills.skills.controllers;
 
 import com.skills.skills.data.SkillsCategoryRepository;
 import com.skills.skills.data.SkillsRepository;
-import com.skills.skills.data.TagRepository;
 import com.skills.skills.data.UserRepository;
-import com.skills.skills.models.*;
-import com.skills.skills.models.UserProfile;
 import com.skills.skills.models.dto.LoginFormDTO;
 import com.skills.skills.models.dto.RegisterFormDTO;
-import com.skills.skills.models.skill.Skill;
 import com.skills.skills.models.user.User;
+import com.skills.skills.models.user.UserProfile;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.Errors;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
 import javax.validation.Valid;
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Optional;
 
 
@@ -33,9 +31,6 @@ public class AuthenticationController {
 
     @Autowired
     SkillsRepository skillsRepository;
-
-    @Autowired
-    public TagRepository tagRepository;
 
     @Autowired
     public SkillsCategoryRepository skillsCategoryRepository;
@@ -61,37 +56,21 @@ public class AuthenticationController {
         session.setAttribute(userSessionKey, user.getId());
 
     }
+//
+//    @GetMapping
+//    public String displayPageAfterLogin(HttpSession session, Model model) {
+//        User user = getUserFormSession(session);
+//        model.addAttribute("user", user);
+//        return "index";
+//    }
 
-    @GetMapping("/users/profile")
+    @GetMapping("")
     public String displayPageAfterLogin (HttpSession session, Model model) {
         User user = getUserFormSession(session);
         model.addAttribute("user", user);
         model.addAttribute("skills", user.getSkills());
-        model.addAttribute(new Skill());
-        model.addAttribute(new Tag());
-        model.addAttribute("tags", tagRepository.findAll());
-        return "users/profile";
-    }
-
-    @PostMapping("/users/profile")
-        public String displayPageAfterFilter (@RequestParam int tagId, HttpSession session, Model model) {
-        User user = getUserFormSession(session);
-
-        List<Skill> skills;
-        List<Skill> filteredSkills = new ArrayList<>();
-        skills = user.getSkills();
-
-        skills.forEach(skill -> {
-            Tag skillTag = skill.tagName;
-            if (skill.getTagId(skillTag) == tagId) {
-                filteredSkills.add(skill);
-            }
-        });
-        model.addAttribute("user", user);
-        model.addAttribute("skills", filteredSkills);
-        model.addAttribute(new Skill());
-        model.addAttribute("tags", tagRepository.findAll());
-        return "users/profile";
+        model.addAttribute("tag", user.getSkills());
+        return "index";
     }
 
 
@@ -138,7 +117,7 @@ public class AuthenticationController {
         //return "redirect:login";
         //return "users/index";
         //return "register_success";
-        return "redirect:/users/profile";
+        return "redirect:";
 
     }
 
@@ -179,7 +158,7 @@ public class AuthenticationController {
 
         setUserInSession(request.getSession(), theUser);
 
-        return "redirect:/users/profile";
+        return "redirect:";
     }
 
     @GetMapping("/logout")
