+++ src/main/java/com/skills/skills/controllers/UserController.java
@@ -1,8 +1,8 @@
 package com.skills.skills.controllers;
 
 import com.skills.skills.data.UserRepository;
-import com.skills.skills.models.UserProfile;
 import com.skills.skills.models.user.User;
+import com.skills.skills.models.user.UserProfile;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
